Safari-611.1.10/LayoutTests/fast/scrolling/rtl-scrollbars-alternate-body-dir-attr-does-not-update-scrollbar-placement-expected.txt

ACTION : Drag the outermost frame LTR scrollbar.
PASS scroller.scrollTop is not 0
ACTION : Flip 'dir' from 'rtl' to 'ltr'.
ACTION : Drag the outermost frame RTL scrollbar.
PASS scroller.scrollTop is not 0
